+++ src/clua-cartesi.cpp
@@ -92,6 +92,46 @@ static int cartesi_mod_keccak(lua_State *L) {
     return 1;
 }
 
+/// \brief This is the cartesi.keccak() function implementation.
+/// \param L Lua state.
+static int cartesi_mod_hash(lua_State *L) {
+    using namespace cartesi;
+    machine_hasher_type h;
+    machine_hasher_type::hash_type hash;
+    if (lua_gettop(L) > 2) {
+        luaL_argerror(L, 3, "too many arguments");
+    }
+    if (lua_gettop(L) < 1) {
+        luaL_argerror(L, 1, "too few arguments");
+    }
+    if (lua_isinteger(L, 1) != 0) {
+        if (lua_gettop(L) > 1) {
+            luaL_argerror(L, 2, "too many arguments");
+        }
+        uint64_t word = luaL_checkinteger(L, 1);
+        h.begin();
+        // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+        h.add_data(reinterpret_cast<const unsigned char *>(&word), sizeof(word));
+        h.end(hash);
+        // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+        lua_pushlstring(L, reinterpret_cast<const char *>(hash.data()), hash.size());
+        return 1;
+    }
+    h.begin();
+    size_t len1 = 0;
+    const char *hash1 = luaL_checklstring(L, 1, &len1);
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    h.add_data(reinterpret_cast<const unsigned char *>(hash1), len1);
+    size_t len2 = 0;
+    const char *hash2 = luaL_optlstring(L, 2, "", &len2);
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    h.add_data(reinterpret_cast<const unsigned char *>(hash2), len2);
+    h.end(hash);
+    // NOLINTNEXTLINE(cppcoreguidelines-pro-type-reinterpret-cast)
+    lua_pushlstring(L, reinterpret_cast<const char *>(hash.data()), hash.size());
+    return 1;
+}
+
 static int cartesi_mod_tobase64(lua_State *L) try {
     size_t size = 0;
     const char *data = luaL_checklstring(L, 1, &size);
@@ -147,7 +187,10 @@ static int cartesi_mod_new(lua_State *L) try {
 
 /// \brief Contents of the cartesi module table.
 static const auto cartesi_mod = clua_make_luaL_Reg_array({
+    // keccak is only used in cmio-test.lua.
+    // If we find a pure Lua implementation, we can remove keccak altogether.
     {"keccak", cartesi_mod_keccak},
+    {"hash", cartesi_mod_hash},
     {"tobase64", cartesi_mod_tobase64},
     {"frombase64", cartesi_mod_frombase64},
     {"tojson", cartesi_mod_tojson},