+++ src/complete-merkle-tree.h
@@ -21,7 +21,7 @@
 #include <type_traits>
 #include <vector>
 
-#include "keccak-256-hasher.h"
+#include "machine-hasher.h"
 #include "merkle-tree-proof.h"
 #include "meta.h"
 #include "pristine-merkle-tree.h"
@@ -39,7 +39,7 @@ namespace cartesi {
 class complete_merkle_tree {
 public:
     /// \brief Hasher class.
-    using hasher_type = keccak_256_hasher;
+    using hasher_type = machine_hasher_type;
 
     /// \brief Storage for a hash.
     using hash_type = hasher_type::hash_type;